# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shake Crazy Food Truck API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MenuItem {
  id: number;
  name: string;
  category: string;
  isVeg: boolean;
  price: number;
  description?: string;
  imageUrl?: string | null;
  isAvailable: boolean;
  updatedAt: string;
}

export interface MenuItemRequest {
  name: string;
  category: string;
  isVeg: boolean;
  price: number;
  description?: string;
  imageUrl?: string | null;
  isAvailable: boolean;
}

export interface Discount {
  id: number;
  name: string;
  discountPercentage: number;
  isActive: boolean;
  isAuto: boolean;
  startDate?: string | null;
  endDate?: string | null;
  posterUrl?: string | null;
  createdAt: string;
}

export interface DiscountRequest {
  name: string;
  discountPercentage: number;
  isActive: boolean;
  isAuto: boolean;
  startDate?: string | null;
  endDate?: string | null;
  posterUrl?: string | null;
}

export interface StopwatchAttemptRequest {
  userName: string;
  timeStopped: number;
  ipAddress?: string;
}

export interface StopwatchAttemptResult {
  isWinner: boolean;
  timeStopped: number;
  message: string;
  alreadyPlayed: boolean;
}

export interface StopwatchWinner {
  id: number;
  userName: string;
  timeStopped: number;
  prize: string;
  createdAt: string;
}

export interface CanPlayStatus {
  canPlay: boolean;
  message: string;
}

export interface Submission {
  id: number;
  userName: string;
  imageUrl?: string | null;
  experienceText: string;
  approved: boolean;
  location?: string | null;
  createdAt: string;
}

export interface SubmissionRequest {
  userName: string;
  imageUrl?: string | null;
  experienceText: string;
  location?: string | null;
}

export interface ApproveSubmissionRequest {
  approved: boolean;
}

export interface TruckLocation {
  id: number;
  currentLocation: string;
  latitude?: number | null;
  longitude?: number | null;
  updatedAt: string;
  updatedBy?: string | null;
}

export interface TruckLocationRequest {
  currentLocation: string;
  latitude?: number | null;
  longitude?: number | null;
}

export interface AnalyticsTrackRequest {
  visitorIp: string;
  country?: string | null;
  city?: string | null;
  userAgent?: string | null;
  page?: string | null;
}

export interface CityCount {
  city: string;
  count: number;
}

export interface VisitRecord {
  visitorIp: string;
  country?: string | null;
  city?: string | null;
  visitedAt: string;
  page?: string | null;
}

export interface AnalyticsData {
  totalVisits: number;
  todayVisits: number;
  topCities: CityCount[];
  recentVisits: VisitRecord[];
}

export interface AdminLoginRequest {
  email: string;
  password: string;
}

export interface AdminLoginResponse {
  token: string;
  email: string;
}

export interface AdminStats {
  totalVisits: number;
  totalMenuItems: number;
  totalWinners: number;
  totalSubmissions: number;
  pendingSubmissions: number;
  activeDiscounts: number;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetMenuItemsParams = {
  category?: string;
  available_only?: boolean;
};

export type GetStopwatchWinnersParams = {
  limit?: number;
};

export type CanPlayStopwatchParams = {
  ip?: string;
};