# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shake Crazy Food Truck API
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItem {
  id: number;
  name: string;
  category: string;
  isVeg: boolean;
  price: number;
  description?: string;
  imageUrl?: string | null;
  isAvailable: boolean;
  updatedAt: string;
}