# lib/api-zod/src/generated/types/stopwatchWinner.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shake Crazy Food Truck API
 * OpenAPI spec version: 0.1.0
 */

export interface StopwatchWinner {
  id: number;
  userName: string;
  timeStopped: number;
  prize: string;
  createdAt: string;
}