# lib/api-zod/src/generated/types/truckLocation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shake Crazy Food Truck API
 * OpenAPI spec version: 0.1.0
 */

export interface TruckLocation {
  id: number;
  currentLocation: string;
  latitude?: number | null;
  longitude?: number | null;
  updatedAt: string;
  updatedBy?: string | null;
}